fix(2_0_1_7): widen CTJoinPacket accountId to ulong (stream server)
Same uint -> ulong shift as X2EnterWorldPacket (a61e03f4), this time
on the stream-protocol side. 2.0.1.7 CTJoinPacket wire layout is:

  accountId  ulong   (8 bytes)
  cookie     uint    (4 bytes)

Matches the 12-byte body the client sends on the stream connection:

  0E 00 01 00 | 01 00 00 00 00 00 00 00 | C9 82 66 9A
  len  opcode | -------- accountId ----- | -- cookie --

Vanilla read accountId as uint, so cookie consumed the upper 4 bytes
(= 0 for any sane account id). StreamManager.Login rejected token = 0,
the stream server dropped the socket, and the client popped a
'Packet Error' a moment after receiving TCJoinResponsePacket.

Read accountId as UInt64, cast back to uint for StreamManager.Login
so its signature stays compatible with master.

+++ AAEmu.Game/Core/Packets/C2S/CTJoinPacket.cs
@@ -8,9 +8,15 @@ public class CTJoinPacket() : StreamPacket(CTOffsets.CTJoinPacket)
 {
     public override void Read(PacketStream stream)
     {
-        var accountId = stream.ReadUInt32();
+        // Same uint -> ulong shift as X2EnterWorldPacket (a61e03f4), this time
+        // on the stream protocol. 2.0.1.7 sends accountId as ulong (8 bytes)
+        // before the uint cookie. Vanilla read 4 bytes, so cookie picked up the
+        // upper 4 bytes of accountId (= 0 for any reasonable account id) and
+        // StreamManager.Login rejected the connection — the stream server then
+        // dropped the socket and the client popped 'Packet Error'.
+        var accountId = stream.ReadUInt64();
         var cookie = stream.ReadUInt32();
 
-        StreamManager.Instance.Login(Connection, accountId, cookie);
+        StreamManager.Instance.Login(Connection, (uint)accountId, cookie);
     }
 }
